--- a/src/date/support-detection/index.js
+++ b/src/date/support-detection/index.js
@@ -1,2 +1,2 @@
-export { isToLocaleDateStringSupported } from './toLocaleDateString';
+export { isIntlSupported } from './intl';
 export { isSelectedLocaleSupported } from './selectedLocale';