… src/date/support-detection/toLocaleDateString.js
… src/date/support-detection/toLocaleDateString.spec.js