Replaces `Date.toLocaleDateString()` with `Intl.DateTimeFormat` for better [performance](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toLocaleDateString#Performance)

**Background**

[Date lookup](https://github.com/transferwise/components/pull/94) uses `formatDate` a lot and I noticed slow rendering speed on low-end devices.

It was caused by `Date.toLocaleDateString()`. `toLocaleDateString` seems to be using `Intl.DateTimeFormat` underneath anyway, but it creates new instance of formatter (`new Intl.DateTimeFormat`) every time, which is slow.

I switched to `Intl.DateTimeFormat` and I'm caching formatters.

**Speed**

Tests on Chrome with 6xCPU slowdown:
```
[...Array(1000)].forEach(() => { formatDate(new Date(), 'en-GB', opts); });
// before: 6739ms
// after:  32ms

dateLookup initial day-calendar render (clicked and opened)
// before: 635ms
// after:  202ms
```